Moving Agent manager to his own folder
